refactor: db層のユニットテストをsrc-tauri/testsに移動する
src-tauri/src/db/schema.rs と repository.rs にあったインラインテストを
CLAUDE.mdの規約に従い src-tauri/tests/db/ 以下に分離する

- tests/db/schema.rs: スキーママイグレーションテスト
- tests/db/repository.rs: リポジトリCRUDテスト
- tests/db/mod.rs: モジュール宣言
- tests/lib.rs: mod db を追加
- src/db/schema.rs: #[cfg(test)]ブロックと未使用定数を削除
- src/db/repository.rs: #[cfg(test)]ブロックを削除
