+++ src-tauri/tests/db/mod.rs
@@ -1,0 +1,2 @@
+mod repository;
+mod schema;
