+++ src-tauri/tests/lib.rs
@@ -1,2 +1,3 @@
 mod api;
 mod common;
+mod db;
